Add files via upload
Having issue with single_plot_multiple_subplots.py. When user checks off plot_types in this file, they should be able to check them off in any order they want, but the code is not yet ready to allow for that. This is where they code has hit a major hitch. 

single_plot_multiple subplots_test.py is to be copied and pasted into the terminal and run in there. The first line is to be re-run with a different number for the subplot to test each if statement inside the for loop.
First: run everything as is. It should run through elif
Second: run the first line while changing subplot number from 0 to 1 or whatever and then run the entire for loop. It should run through if ONLY ONCE!!! It runs through it twice which is the current issue I am having. I tried using a break, but then it cannot compare all the keys
Lastly: run the first line again while changing the subplot number back to 0: It should run through else, AGAIN ONLY ONCE!!! I cannot get it to run through else at all. This is the most important thing to get to work as it is where the user picks a plot type for a previous subplot.