Add files via upload
Added add() function and made the file iteratively testable. Made the program take into account that the arrays appended to are fundamentally different on the first set of loops where subplot value is the same each time. This was taken into account by adding if else statements inside of the if else statements. Also forgot to pop keys in the else: statement at the end of the program. Now successfully allocates subplots = [0, 0, 0] and will now need to work on allocating when subplot value changes to something other than 0.