Add files via upload
Finally finished the single_plot_multiple_subplots file with one minor exception in the form of an error at the final step before plotting. Not sure what is causing it, but the exact sequence to reproduce the error is rather random. I will try to find a solution later, but for now, I think it is in a good enough state for some initial test running.